multimodal_vqa/requirements.txt
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.20.0
datasets>=2.0.0
pillow>=8.0.0
nltk>=3.6.0
scikit-learn>=1.0.0
numpy>=1.20.0
safetensors>=0.3.0
tqdm>=4.60.0
pandas>=1.3.0
streamlit>=1.15.0
accelerate>=0.26.0
